Update User.js
I added an if statement in line 35 of the function create_account() to handle an edge case. If the number of users for our application is 0 and this function is called an error condition was formed when doing 'data3 = Object.keys(data2).length' as data2 is null as there are no users in the database.

I have also removed the 'messages' field from the user account.