Update Tutor.js
Redid the loops for the preferred days and timings to make them single loops instead of double loops in create_profile(). 
Thus, had to also redesign the update_preferred_days() and update_preferred_timings() in terms of their loops.
Removed the message field from the tutor profile. Added the 'SessionsHosted' field and the 'Badges' field to the tutor account.
Created the update_badges_for_tutor() so that I can set the required badges as true. I plan to call this function every time a session has been completed (this part will be implemented later).