Update Requests.js
Made the following changes to the request class

create_request
Removed the field 'TutorsWhoAccepted' with the field 'Offers'.
I have an additional field 'MatchedTutors' but its main purpose to help me in the backend to delete the request from all tutors who were matched from the matching algorithm. The Language field of the request is set to the user's preferred language

delete_request
minor changes made to one of the for loops but it will not impact the front end.

add_tutor_to_request()
this function will ask for additional parameters so that the offer created by the tutor for that request can be saved in the db. I will now require inputs in the following order -- requestID, tutorID, startTime, length, date, location, format.
This is how the offers will be stored in the db
Requests
      |
      ---- RequestID
                   |
                   ---- Offers
                              |
                              --- 0 : 'Offer ID'
                              |
                              --- 1
                                     |
                                     --- Date : ....
                                     --- Location: ...
                                     --- Time: ...
                                     --- Length: ...
                                     --- TutorID: ...
                                     --- id: length of offers field - 1

reject_request()
I added additional for loop but it has no impact on front end. 

I have REMOVED remove_tutor_from_request()