Update Tutor.js
Made changes to the delete_profile() function. It will now delete all the data associated with that tutor account

The create_profile function -- I use the list of courses the tutor wants to tutor for to update the number of tutors who tutor that course and add the course if it is not in the db.
Example1, for CS180 there are 2 tutors. A new tutor wants to tutor for CS180, thus the count increases from 2 to 3
Example2, if CS242 is not the db and a tutor wants to tutor for it, I add the course as follows CS242: 1, 1 refers to the number of tutors for that course
Example3, if the number of tutors for CS373 was 1 and the tutor leaves (the count becomes 0), the course is removed