Update Requests.js
For the function add_tutor_to_request() will save the offer made by the tutor as Requests/requestID/Offers/tutorID instead of Requests/requestID/Offers/Offer n, where n is the number of offers for the given request. I did this to handle an edge case in which adding a new offer may overwrite the contents of a previous offer.